Save button belongs in BlockWriter, not BlockReader.
