+++ client/src/main/typescript/components/BlockReader/BlockReader.tsx
@@ -1,8 +1,6 @@
 import * as React from "react";
 import * as ReactDOM from "react-dom";
 import BlockEditor, { TakeDocument } from "../BlockEditor";
-import { DraftRev } from "../../java2ts/DraftRev";
-import { DraftPost } from "../../java2ts/DraftPost";
 
 interface BlockReaderProps {
   initState: TakeDocument;
@@ -10,7 +8,6 @@ interface BlockReaderProps {
 
 interface BlockReaderState {
   takeDocument: TakeDocument;
-  parentRev?: DraftRev;
 }
 
 class BlockReader extends React.Component<BlockReaderProps, BlockReaderState> {
@@ -43,54 +40,12 @@ class BlockReader extends React.Component<BlockReaderProps, BlockReaderState> {
       "&" +
       offset;
   };
-  handleSaveClick = () => {
-    const headers = new Headers();
-
-    headers.append("Accept", "application/json"); // This one is enough for GET requests
-    headers.append("Content-Type", "application/json"); // This one sends body
-
-    //TODO: enforce title length <= 255
-    const bodyJson: DraftPost = {
-      parentRev: this.state.parentRev,
-      title: this.state.takeDocument.title,
-      blocks: this.state.takeDocument.blocks
-    };
-    const request: RequestInit = {
-      method: "POST",
-      credentials: "include",
-      headers: headers,
-      body: JSON.stringify(bodyJson)
-    };
-    fetch("/drafts/save", request)
-      .then(function(response) {
-        const contentType = response.headers.get("content-type");
-        if (contentType && contentType.indexOf("application/json") >= 0) {
-          return response.json();
-        }
-        throw new TypeError("Oops, we haven't got JSON!");
-      })
-      .then(function(json) {
-        const parentRev: DraftRev = json;
-        this.setState({
-          parentRev: parentRev
-        });
-      })
-      .catch(function(error) {
-        throw error;
-      });
-  };
   render() {
     const eventHandlers = {
       onDocumentClick: this.handleClick
     };
     return (
       <div>
-        <button
-          className="editor__button video__button--save"
-          onClick={this.handleSaveClick}
-        >
-          Save
-        </button>
         <BlockEditor
           takeDocument={(Object as any).assign({}, this.state.takeDocument)}
           eventHandlers={eventHandlers}
